Update jobacctinfo struct with 64 bit user_cpu_sec and sys_cpu_sec.
Update pack/unpack to handle this correctly.  For old protocol, send
NO_VAL if the number is too large.

Bug 10903

+++ src/common/slurm_jobacct_gather.h
@@ -80,9 +80,9 @@ typedef struct {
 
 struct jobacctinfo {
 	pid_t pid;
-	uint32_t sys_cpu_sec;
+	uint64_t sys_cpu_sec;
 	uint32_t sys_cpu_usec;
-	uint32_t user_cpu_sec;
+	uint64_t user_cpu_sec;
 	uint32_t user_cpu_usec;
 	uint32_t act_cpufreq; /* actual cpu frequency */
 	acct_gather_energy_t energy;
